requirements.txt
streamlit>=1.28.0
psycopg2-binary>=2.9.9
spacy>=3.7.0
python-dotenv>=1.0.0
sentence-transformers>=2.2.0
pytest>=7.4.0

# Phase 2: Resume Parsing
pdfplumber>=0.9.0
python-docx>=0.8.11

# Phase 2: Analysis
scikit-learn>=1.0.0

# LLM Verification
openai>=1.0.0
